删除未使用过的命令行工具类CommandUtils，java压缩工具类DeflateUtils，springboot不再需要的配置文件工具类PropertiesUtils，可被BusinessIdUtils替换的唯一id生成器OnlyIdGenerator；web.*不再推荐使用;

+++ src/main/java/cn/emay/utils/number/IntegerUtils.java
@@ -2,35 +2,35 @@ package cn.emay.utils.number;
 
 /**
  * int工具类
- * 
- * @author Frank
  *
+ * @author Frank
  */
 public class IntegerUtils {
 
-	/**
-	 * int 到4位字节
-	 * 
-	 * @param value
-	 * @return
-	 */
-	public static byte[] toBytes(int value) {
-		byte[] src = new byte[4];
-		src[0] = (byte) ((value >> 24) & 0xFF);
-		src[1] = (byte) ((value >> 16) & 0xFF);
-		src[2] = (byte) ((value >> 8) & 0xFF);
-		src[3] = (byte) (value & 0xFF);
-		return src;
-	}
+    /**
+     * int 到4位字节
+     *
+     * @param value 数字
+     * @return 字节
+     */
+    public static byte[] toBytes(int value) {
+        byte[] src = new byte[4];
+        src[0] = (byte) ((value >> 24) & 0xFF);
+        src[1] = (byte) ((value >> 16) & 0xFF);
+        src[2] = (byte) ((value >> 8) & 0xFF);
+        src[3] = (byte) (value & 0xFF);
+        return src;
+    }
 
-	/**
-	 * 字节转int
-	 * 
-	 * @param src
-	 * @return
-	 */
-	public static int toInt(byte[] src) {
-		return (int) (((src[0] & 0xFF) << 24) | ((src[1] & 0xFF) << 16) | ((src[2] & 0xFF) << 8) | (src[3] & 0xFF));
-	}
+    /**
+     * 字节转int
+     *
+     * @param src 字节
+     * @return 数字
+     */
+    public static int toInt(byte[] src) {
+        return ((src[0] & 0xFF) << 24) | ((src[1] & 0xFF) << 16) | ((src[2] & 0xFF) << 8) | (src[3] & 0xFF);
+    }
+
 
 }
